interfaces/audio-playback: let pulseaudio own org.pulseaudio.Server
Let pulseaudio own the org.pulseaudio.Server dBus name. It is used for
service tracking purposes and it is needed when pulseaudio is compiled
with dBus support, which in turn is needed to access bluez and be able
to handle bluetooth sound devices. Note that this is unrelated to the
not maintained org.PulseAudio.Core1 dBus interface [1].

[1] https://www.freedesktop.org/wiki/Software/PulseAudio/Documentation/Developer/Clients/DBus/