boot: do not use try logic when reverting to a previous base
Previously we were using try/trying logic for bases when reverting to
a previous base in the case of having boot already to the base to be
reverted. This caused a boot loop as the new base was not present
after rebooting and the snap-bootstrap logic checks that. Prevent that
by directly reverting to the previous base, which we already now is
safe, and it would not make sense to "revert" to the base we are
reverting.